En egen komponent for kvitterinsSiden (#77)
### Hvorfor er dette nødvendig? ✨

For at kvittering komponent skal brukes på alle kvitteringsider til alle
stønader så skal det være en egen komponent
[Lenke til trello
kort](https://trello.com/c/hmqia0Zb/123-eget-komponent-for-kvittering)

### Hvordan er det løst? 🧠

Mappen sider er lagt til i app og det er laget en ny komponent
kvitteringSide.tsx i denne mappen og inneholder det man trenger for en
kvittering. KvitteringsSide ble laget som et eget komponent så den kan
brukes på alle stønader.
